users: move wait list & room user list containers to `containers/`
