+++ src/components/App/index.js
@@ -3,8 +3,8 @@ import { DragDropContext } from 'react-dnd';
 import HTML5Backend from 'react-dnd-html5-backend';
 import Chat from '../../containers/Chat';
 import ChatInput from '../Chat/Input';
-import RoomUserList from '../UserList/RoomUserList';
-import WaitList from '../UserList/WaitList';
+import RoomUserList from '../../containers/RoomUserList';
+import WaitList from '../../containers/WaitList';
 import FooterBar from '../../containers/FooterBar';
 import HeaderBar from '../../containers/HeaderBar';
 import PanelSwitcher from '../PanelSwitcher';
