+++ src/containers/WaitList.js
@@ -1,12 +1,13 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import UserList from '../UserList';
+import { createStructuredSelector } from 'reselect';
 
-function mapStateToProps({ waitlist, users }) {
-  return {
-    users: waitlist.waitlist.map(id => users[id])
-  };
-}
+import { waitlistUsersSelector } from '../selectors/waitlistSelectors';
+import UserList from '../components/UserList';
+
+const mapStateToProps = createStructuredSelector({
+  users: waitlistUsersSelector
+});
 
 @connect(mapStateToProps)
 export default class WaitListContainer extends Component {
